backend/requirements.txt
alembic==1.4.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
midtransclient==1.1.1
mysqlclient==2.0.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
Werkzeug==1.0.1
